Integer id -> Long id; Account has a mother; Transactions are made of mutiple Entries

+++ api/src/main/java/br/com/pastaeriso/api/generics/address/AddressRepository.java
@@ -6,6 +6,6 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 
 @Repository
 @CrossOrigin
-public interface AddressRepository extends JpaRepository<Address, Integer> {
+public interface AddressRepository extends JpaRepository<Address, Long> {
 
 }
