+++ api/src/main/java/br/com/pastaeriso/api/purchases/purchase/products/PurchaseProductRepository.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 
 @Repository
 @CrossOrigin
-public interface PurchaseProductRepository extends JpaRepository<PurchaseProduct, Integer>{
+public interface PurchaseProductRepository extends JpaRepository<PurchaseProduct, Long>{
 
 	public List<PurchaseProduct> findByProductName(String productName);
 	public List<PurchaseProduct> findByProductNameIgnoreCase(String productName);
