+++ api/src/main/java/br/com/pastaeriso/api/sales/delivery/deliveryman/contract/ContractRepository.java
@@ -6,6 +6,6 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 
 @Repository
 @CrossOrigin
-public interface ContractRepository extends JpaRepository<Contract, Integer> {
+public interface ContractRepository extends JpaRepository<Contract, Long> {
 
 }
